\docType{data}
\name{big_portfolio}
\alias{big_portfolio}
\title{Big Portfolio}
\format{A data frame with 10000 rows 6 variables.}
\description{
Sample dataset.
}
\details{
\itemize{ \item id \item origination_date \item balance
\item product \item region \item repayment_date }
}
\keyword{datasets}

