\name{get_calculated_member_xml}
\alias{get_calculated_member_xml}
\title{Get XML for calculated member}
\usage{
  get_calculated_member_xml(engine, calculated_member = NA,
    debug = FALSE)
}
\arguments{
  \item{engine}{See \code{\link{create_schema}} for
  details.}

  \item{calculated_member}{See \code{\link{create_schema}}
  for details.}

  \item{debug}{See \code{\link{create_schema}} for
  details.}
}
\description{
  Returns XML for calculated members.
}

