\name{is_vector}
\alias{is_vector}
\title{Check if argument is vector with given properties}
\usage{
  is_vector(vector, length = 1, mode = "character")
}
\arguments{
  \item{vector}{Input object.}

  \item{length}{Required length of vector.}

  \item{mode}{Required mode of vector}
}
\description{
  Checks if argument is vector with given length and mode.
  Used to control arguments. Returns error when object does
  not match given properties.
}

