\name{standardize_name}
\alias{standardize_name}
\title{Standardize name}
\usage{
  standardize_name(name = NA)
}
\arguments{
  \item{Name}{String to be standardized.}
}
\description{
  Clean name into more human-readable form. Removes
  undescores, makes first letter upper-case.
}
\examples{
standardize_name("column_name")
}

